+++ R/hardhat.R
@@ -47,8 +47,9 @@
 #'
 #' @section Multi-outcome:
 #'
-#' TabNet allows multi-outcome prediction, the outcomes must all be numeric or all be categorical,
-#' and the __data frame__ method shall be used for `x` and `y`.
+#' TabNet allows multi-outcome prediction, which is usually named [multi-label classification](https://en.wikipedia.org/wiki/Multi-label_classification)
+#'   or multi-output classification when outcomes are categorical.
+#' Multi-outcome currently expect outcomes to be either all numeric or all categorical.
 #'
 #' @section Threading:
 #'
@@ -62,19 +63,37 @@
 #' torch::torch_set_num_interop_threads(1)
 #' ```
 #'
-#' @examples
-#' if (torch::torch_is_installed()) {
+#' @examplesIf torch::torch_is_installed()
 #'
-#' # regression using formula specification
 #' data("ames", package = "modeldata")
+#' data("attrition", package = "modeldata")
+#' ids <- sample(nrow(attrition), 256)
+#'
+#' ## Single-outcome regression using formula specification
 #' fit <- tabnet_fit(Sale_Price ~ ., data = ames, epochs = 1)
 #'
-#' # classification using data-frame specification
-#' data("attrition", package = "modeldata")
+#' ## Single-outcome classification using data-frame specification
 #' attrition_x <- attrition[,-which(names(attrition) == "Attrition")]
-#' fit <- tabnet_fit(attrition_x, attrition$Attrition, epochs = 1)
-#' }
+#' fit <- tabnet_fit(attrition_x, attrition$Attrition, epochs = 1, verbose = TRUE)
 #'
+#' ## Multi-outcome regression on `Sale_Price` and `Pool_Area` in `ames` dataset using formula,
+#' ames_fit <- tabnet_fit(Sale_Price + Pool_Area ~ ., data = ames[ids,], epochs = 2, valid_split = 0.2)
+#'
+#' ## Multi-label classification on `Attrition` and `JobSatisfaction` in
+#' ## `attrition` dataset using recipe
+#' library(recipes)
+#' rec <- recipe(Attrition + JobSatisfaction ~ ., data = attrition[ids,]) %>%
+#'   step_normalize(all_numeric(), -all_outcomes())
+#'
+#' attrition_fit <- tabnet_fit(rec, data = attrition[ids,], epochs = 2, valid_split = 0.2)
+#'
+#' ## Hierarchical classification on  `acme`
+#' data(acme, package = "data.tree")
+#'
+#' acme_fit <- tabnet_fit(acme, epochs = 2, verbose = TRUE)
+#'
+#' # Note: Dataset number of rows and model number of epochs should be increased
+#' # for publication-level results.
 #' @return A TabNet model object. It can be used for serialization, predictions, or further fitting.
 #'
 #' @export
@@ -85,10 +104,9 @@ tabnet_fit <- function(x, ...) {
 #' @export
 #' @rdname tabnet_fit
 tabnet_fit.default <- function(x, ...) {
-  stop(
-    "`tabnet_fit()` is not defined for a '", class(x)[1], "'.",
-    call. = FALSE
-  )
+  stop(domain=NA,
+       gettextf("`tabnet_fit()` is not defined for a '%s'.", class(x)[1]),
+       call. = FALSE)
 }
 
 #' @export
@@ -107,7 +125,7 @@ tabnet_fit.data.frame <- function(x, y, tabnet_model = NULL, config = tabnet_con
     ]
   config <- utils::modifyList(config, as.list(new_config))
 
-  tabnet_bridge(processed, config = config, tabnet_model, from_epoch, task="supervised")
+  tabnet_bridge(processed, config = config, tabnet_model, from_epoch, task = "supervised")
 }
 
 #' @export
@@ -132,7 +150,7 @@ tabnet_fit.formula <- function(formula, data, tabnet_model = NULL, config = tabn
   ]
   config <- utils::modifyList(config, as.list(new_config))
 
-  tabnet_bridge(processed, config = config, tabnet_model, from_epoch, task="supervised")
+  tabnet_bridge(processed, config = config, tabnet_model, from_epoch, task = "supervised")
 }
 
 #' @export
@@ -151,7 +169,41 @@ tabnet_fit.recipe <- function(x, data, tabnet_model = NULL, config = tabnet_conf
   ]
   config <- utils::modifyList(config, as.list(new_config))
 
-  tabnet_bridge(processed, config = config, tabnet_model, from_epoch, task="supervised")
+  tabnet_bridge(processed, config = config, tabnet_model, from_epoch, task = "supervised")
+}
+
+#' @export
+#' @rdname tabnet_fit
+#'
+#' @importFrom dplyr filter mutate select mutate_all mutate_if
+#' @importFrom tidyr replace_na
+#'
+tabnet_fit.Node <- function(x, tabnet_model = NULL, config = tabnet_config(), ..., from_epoch = NULL) {
+  # ensure there is no level_* col in the Node object
+  check_compliant_node(x)
+  # get tree leaves and extract attributes into data.frames
+  xy_df <- node_to_df(x)
+  processed <- hardhat::mold(xy_df$x, xy_df$y)
+  # Given n classes, M is an (n x n) matrix where M_ij = 1 if class i is descendant of class j
+  ancestor <- data.tree::ToDataFrameNetwork(x) %>%
+   mutate_if(is.character, ~.x %>% as.factor %>% as.numeric)
+  # TODO check correctness
+  # embed the M matrix in the config$ancestor variable
+  dims <- c(max(ancestor), max(ancestor))
+  ancestor_m <- Matrix::sparseMatrix(ancestor$from, ancestor$to, dims = dims, x = 1)
+  check_type(processed$outcomes)
+
+  default_config <- tabnet_config()
+  new_config <- do.call(tabnet_config, list(...))
+  new_config <- new_config[
+    mapply(
+      function(x, y) ifelse(is.null(x), !is.null(y), x != y),
+      default_config,
+      new_config)
+  ]
+  config <- utils::modifyList(config, as.list(new_config, ancestor = ancestor_m))
+
+  tabnet_bridge(processed, config = config, tabnet_model, from_epoch, task = "supervised")
 }
 
 new_tabnet_fit <- function(fit, blueprint) {
@@ -226,11 +278,9 @@ new_tabnet_fit <- function(fit, blueprint) {
 #' torch::torch_set_num_interop_threads(1)
 #' ```
 #'
-#' @examples
-#' if (torch::torch_is_installed()) {
+#' @examplesIf torch::torch_is_installed()
 #' data("ames", package = "modeldata")
 #' pretrained <- tabnet_pretrain(Sale_Price ~ ., data = ames, epochs = 1)
-#' }
 #'
 #' @return A TabNet model object. It can be used for serialization, predictions, or further fitting.
 #'
@@ -242,10 +292,9 @@ tabnet_pretrain <- function(x, ...) {
 #' @export
 #' @rdname tabnet_pretrain
 tabnet_pretrain.default <- function(x, ...) {
-  stop(
-    "`tabnet_pretrain()` is not defined for a '", class(x)[1], "'.",
-    call. = FALSE
-  )
+  stop(domain=NA,
+       gettextf("`tabnet_pretrain()` is not defined for a '%s'.", class(x)[1]),
+       call. = FALSE)
 }
 
 
@@ -265,7 +314,7 @@ tabnet_pretrain.data.frame <- function(x, y, tabnet_model = NULL, config = tabne
   ]
   config <- utils::modifyList(config, as.list(new_config))
 
-  tabnet_bridge(processed, config = config, tabnet_model, from_epoch, task="unsupervised")
+  tabnet_bridge(processed, config = config, tabnet_model, from_epoch, task = "unsupervised")
 }
 
 #' @export
@@ -289,7 +338,7 @@ tabnet_pretrain.formula <- function(formula, data, tabnet_model = NULL, config =
   ]
   config <- utils::modifyList(config, as.list(new_config))
 
-  tabnet_bridge(processed, config = config, tabnet_model, from_epoch, task="unsupervised")
+  tabnet_bridge(processed, config = config, tabnet_model, from_epoch, task = "unsupervised")
 }
 
 #' @export
@@ -307,7 +356,18 @@ tabnet_pretrain.recipe <- function(x, data, tabnet_model = NULL, config = tabnet
   ]
   config <- utils::modifyList(config, as.list(new_config))
 
-  tabnet_bridge(processed, config = config, tabnet_model, from_epoch, task="unsupervised")
+  tabnet_bridge(processed, config = config, tabnet_model, from_epoch, task = "unsupervised")
+}
+
+#' @export
+#' @rdname tabnet_pretrain
+tabnet_pretrain.Node <- function(x, tabnet_model = NULL, config = tabnet_config(), ..., from_epoch = NULL) {
+  # ensure there is no level_* col in the Node object
+  check_compliant_node(x)
+  # get tree leaves and extract attributes into data.frames
+  xy_df <- node_to_df(x)
+  tabnet_pretrain(xy_df$x, xy_df$y, tabnet_model = tabnet_model, config = config, ..., from_epoch = from_epoch)
+
 }
 
 new_tabnet_pretrain <- function(pretrain, blueprint) {
@@ -326,14 +386,16 @@ tabnet_bridge <- function(processed, config = tabnet_config(), tabnet_model, fro
   predictors <- processed$predictors
   outcomes <- processed$outcomes
   epoch_shift <- 0L
+
   if (!(is.null(tabnet_model) || inherits(tabnet_model, "tabnet_fit") || inherits(tabnet_model, "tabnet_pretrain")))
-    rlang::abort(paste0(tabnet_model," is not recognised as a proper TabNet model"))
+    stop(gettextf("'%s' is not recognised as a proper TabNet model", tabnet_model),
+         call. = FALSE)
 
   if (!is.null(from_epoch) && !is.null(tabnet_model)) {
     # model must be loaded from checkpoint
 
     if (from_epoch > (length(tabnet_model$fit$checkpoints) * tabnet_model$fit$config$checkpoint_epoch))
-      rlang::abort(paste0("The model was trained for less than ", from_epoch, " epochs"))
+      stop(gettextf("The model was trained for less than '%s' epochs", from_epoch), call. = FALSE)
 
     # find closest checkpoint for that epoch
     closest_checkpoint <- from_epoch %/% tabnet_model$fit$config$checkpoint_epoch
@@ -345,7 +407,7 @@ tabnet_bridge <- function(processed, config = tabnet_config(), tabnet_model, fro
   }
   if (task == "supervised") {
     if (sum(is.na(outcomes)) > 0) {
-      rlang::abort("Error: found missing values in the outcome data.")
+      stop(gettextf("Found missing values in the `%s` outcome column.", names(outcomes)), call. = FALSE)
     }
     if (is.null(tabnet_model)) {
       # new supervised model needs network initialization
@@ -355,7 +417,7 @@ tabnet_bridge <- function(processed, config = tabnet_config(), tabnet_model, fro
     } else if (!check_net_is_empty_ptr(tabnet_model) && inherits(tabnet_model, "tabnet_fit")) {
       # resume training from supervised
       if (!identical(processed$blueprint, tabnet_model$blueprint))
-        rlang::abort("Model dimensions don't match.")
+        stop("Model dimensions don't match.", call. = FALSE)
 
       # model is available from tabnet_model$serialized_net
       m <- reload_model(tabnet_model$serialized_net)
@@ -380,7 +442,7 @@ tabnet_bridge <- function(processed, config = tabnet_config(), tabnet_model, fro
       tabnet_model$fit$network <- reload_model(tabnet_model$fit$checkpoints[[last_checkpoint]])
       epoch_shift <- last_checkpoint * tabnet_model$fit$config$checkpoint_epoch
 
-    } else rlang::abort(paste0("No model serialized weight can be found in ", tabnet_model, ", check the model history"))
+    } else stop(gettextf("No model serialized weight can be found in `%s`, check the model history", tabnet_model), call. = FALSE)
 
     fit_lst <- tabnet_train_supervised(tabnet_model, predictors, outcomes, config = config, epoch_shift)
     return(new_tabnet_fit(fit_lst, blueprint = processed$blueprint))
@@ -397,57 +459,18 @@ tabnet_bridge <- function(processed, config = tabnet_config(), tabnet_model, fro
 #' @importFrom stats predict
 #' @export
 predict.tabnet_fit <- function(object, new_data, type = NULL, ..., epoch = NULL) {
+  if (inherits(new_data, "Node")) {
+    new_data_df <- node_to_df(new_data)$x
+  } else {
+    new_data_df <- new_data
+  }
   # Enforces column order, type, column names, etc
-  processed <- hardhat::forge(new_data, object$blueprint)
+  processed <- hardhat::forge(new_data_df, object$blueprint)
   batch_size <- object$fit$config$batch_size
   out <- predict_tabnet_bridge(type, object, processed$predictors, epoch, batch_size)
-  hardhat::validate_prediction_size(out, new_data)
+  hardhat::validate_prediction_size(out, new_data_df)
   out
 }
-
-#' Check consistency between modeling-task type and class of outcomes vars.
-#'
-#' infer default modeling-task type from the outcome vars class if needed.
-#'
-#' @param outcome_ptype shall be `model$blueprint$ptypes$outcomes` when called from
-#'  a model object, or `processed$outcomes` from the result of a `mold()`
-#' @param type expected type within  `c("numeric", "prob", "class")`
-#'
-#' @return valid type within `c("numeric", "prob", "class")` for repectively regression,
-#' class probabilities, or classification
-#' @noRd
-check_type <- function(outcome_ptype, type = NULL) {
-
-  # outcome_ptype <- model$blueprint$ptypes$outcomes when called from model
-  outcome_all_factor <- all(purrr::map_lgl(outcome_ptype, is.factor))
-  outcome_all_numeric <- all(purrr::map_lgl(outcome_ptype, is.numeric))
-
-  if (!outcome_all_numeric && !outcome_all_factor)
-    rlang::abort(glue::glue("Mixed multi-outcome type '{unique(purrr::map_chr(outcome_ptype, ~class(.x)[[1]]))}' is not supported"))
-
-  if (is.null(type)) {
-    if (outcome_all_factor)
-      type <- "class"
-    else if (outcome_all_numeric)
-      type <- "numeric"
-    else if (ncol(outcome_ptype) == 1)
-      rlang::abort(glue::glue("Unknown outcome type '{class(outcome_ptype)}'"))
-  }
-
-  type <- rlang::arg_match(type, c("numeric", "prob", "class"))
-
-  if (outcome_all_factor) {
-    if (!type %in% c("prob", "class"))
-      rlang::abort(glue::glue("Outcome is factor and the prediction type is '{type}'."))
-  } else if (outcome_all_numeric) {
-    if (type != "numeric")
-      rlang::abort(glue::glue("Outcome is numeric and the prediction type is '{type}'."))
-  }
-
-  invisible(type)
-}
-
-
 
 predict_tabnet_bridge <- function(type, object, predictors, epoch, batch_size) {
 
@@ -461,7 +484,7 @@ predict_tabnet_bridge <- function(type, object, predictors, epoch, batch_size) {
   if (!is.null(epoch)) {
 
     if (epoch > (length(object$fit$checkpoints) * object$fit$config$checkpoint_epoch))
-      rlang::abort(paste0("The model was trained for less than ", epoch, " epochs"))
+      stop(gettextf("The model was trained for less than `%s` epochs", epoch), call. = FALSE)
 
     # find closest checkpoint for that epoch
     ind <- epoch %/% object$fit$config$checkpoint_epoch
@@ -506,7 +529,7 @@ model_pretrain_to_fit <- function(obj, x, y, config = tabnet_config()) {
   m <- reload_model(obj$serialized_net)
 
   if (m$input_dim != tabnet_model_lst$network$input_dim)
-    rlang::abort("Model dimensions don't match.")
+    stop("Model dimensions don't match.", call. = FALSE)
 
   # perform update of selected weights into new tabnet_model
   m_stat_dict <- m$state_dict()
@@ -538,6 +561,131 @@ is_null_external_pointer <- function(pointer) {
   out <- identical(pointer, methods::new("externalptr"))
   attributes(pointer) <- a
   out
+}
+
+#' Check consistency between modeling-task type and class of outcomes vars.
+#'
+#' infer default modeling-task type from the outcome vars class if needed.
+#'
+#' @param outcome_ptype shall be `model$blueprint$ptypes$outcomes` when called from
+#'  a model object, or `processed$outcomes` from the result of a `mold()`
+#' @param type expected type within  `c("numeric", "prob", "class")`
+#'
+#' @return valid type within `c("numeric", "prob", "class")` for respectively regression,
+#' class probabilities, or classification
+#' @noRd
+check_type <- function(outcome_ptype, type = NULL) {
+
+  # outcome_ptype <- model$blueprint$ptypes$outcomes when called from model
+  outcome_all_factor <- all(purrr::map_lgl(outcome_ptype, is.factor))
+  outcome_all_numeric <- all(purrr::map_lgl(outcome_ptype, is.numeric))
+
+  if (!outcome_all_numeric && !outcome_all_factor)
+    stop(gettextf("Mixed multi-outcome type '%s' is not supported", unique(purrr::map_chr(outcome_ptype, ~class(.x)[[1]]))), call. = FALSE)
+
+  if (is.null(type)) {
+    if (outcome_all_factor)
+      type <- "class"
+    else if (outcome_all_numeric)
+      type <- "numeric"
+    else if (ncol(outcome_ptype) == 1)
+      stop(gettextf("Unknown outcome type '%s'", class(outcome_ptype)), call. = FALSE)
+  }
+
+  type <- rlang::arg_match(type, c("numeric", "prob", "class"))
+
+  if (outcome_all_factor) {
+    if (!type %in% c("prob", "class"))
+      stop(gettextf("Outcome is factor and the prediction type is '%s'.", type), call. = FALSE)
+  } else if (outcome_all_numeric) {
+    if (type != "numeric")
+      stop(gettextf("Outcome is numeric and the prediction type is '%s'.", type), call. = FALSE)
+  }
+
+  invisible(type)
+}
+
+
+#' Check that Node object names are compliant
+#'
+#' @param node the Node object, or a dataframe ready to be parsed by `data.tree::as.Node()`
+#'
+#' @return node if it is compliant, else an Error with the column names to fix
+#' @export
+#'
+#' @examplesIf (require("data.tree") || require("dplyr"))
+#' library(dplyr)
+#' library(data.tree)
+#' data(starwars)
+#' starwars_tree <- starwars %>%
+#'   mutate(pathString = paste("tree", species, homeworld, `name`, sep = "/"))
+#'
+#' # pre as.Node() check
+#' try(check_compliant_node(starwars_tree))
+#'
+#' # post as.Node() check
+#' check_compliant_node(as.Node(starwars_tree))
+#'
+check_compliant_node <- function(node) {
+  #  prevent reserved data.tree Node colnames and the level_1 ... level_n names used for coercion
+  if (inherits(node, "Node")) {
+    # Node has already lost its reserved colnames
+    reserved_names <- paste0("level_", c(1:node$height))
+    actual_names <- node$attributesAll
+  } else if (inherits(node, "data.frame") && "pathString" %in% colnames(node)) {
+    node_height <- max(stringr::str_count(node$pathString, "/"))
+    reserved_names <- c(paste0("level_", c(1:node_height)), data.tree::NODE_RESERVED_NAMES_CONST)
+    actual_names <- colnames(node)[!colnames(node) %in% "pathString"]
+  } else {
+    stop("The provided hierarchical object is not recognized with a valid format that can be checked", call. = FALSE)
+  }
+
+  if (any(actual_names %in% reserved_names)) {
+    stop(domain=NA,
+         gettextf("The attributes or colnames in the provided hierarchical object use the following reserved names : '%s'. Please change those names as they will lead to unexpected tabnet behavior.",
+          paste(actual_names[actual_names %in% reserved_names], collapse = "', '")
+         ),
+         call. = FALSE)
+  }
+
+  invisible(node)
+}
+
+#' Turn a Node object into predictor and outcome.
+#'
+#' @param x Node object
+#' @param drop_last_level TRUE unused
+#'
+#' @return a named list of x and y, being respectively the predictor data-frame and the outcomes data-frame,
+#'   as expected inputs for `hardhat::mold()` function.
+#' @export
+#'
+#' @examplesIf (require("data.tree") || require("dplyr"))
+#' library(dplyr)
+#' library(data.tree)
+#' data(starwars)
+#' starwars_tree <- starwars %>%
+#'   mutate(pathString = paste("tree", species, homeworld, `name`, sep = "/")) %>%
+#'   as.Node()
+#' node_to_df(starwars_tree)$x %>% head()
+#' node_to_df(starwars_tree)$y %>% head()
+#' @importFrom dplyr last_col mutate mutate_if select starts_with where
+node_to_df <- function(x, drop_last_level = TRUE) {
+  # TODO get rid of all those import through base R equivalent
+  xy_df <- data.tree::ToDataFrameTypeCol(x, x$attributesAll)
+  x_df <- xy_df %>%
+    select(-starts_with("level_")) %>%
+    mutate_if(is.character, as.factor)
+  y_df <- xy_df %>%
+    select(starts_with("level_")) %>%
+    # drop first (and all zero-variance) column
+    select(where(~ nlevels(as.factor(.x)) > 1 )) %>%
+    # TODO take the drop_last_level param into account
+    # drop last level column
+    select(-last_col()) %>%
+    # TODO impute "NA" with parent through coalesce() via an option
+    mutate_if(is.character, as.factor)
+  return(list(x = x_df, y = y_df))
 }
 
 reload_model <- function(object) {